+++ src/main/java/nextstep/subway/common/config/JpaConfiguration.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common;
+package nextstep.subway.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
